triptracks_app/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2              # DB config via ENV vars
Django==1.11.16                     # Web framework
django-appconf==1.0.2
django-tastypie==0.14.0
django-extensions==1.8.1            # Various Django extras
django-uuidfield==0.5.0
dj-static==0.0.6
gpxpy==1.2.0
gunicorn==19.7.1
# html==1.16
httplib2==0.11.0
ipython==5.5.0                      # Better Django shell experience
idna==2.6
ipaddress==1.0.19
jsonfield==2.0.2
lxml==4.2.1
oauth2client==4.1.2
psycopg2-binary==2.7.5                    # Postgres connector
pyasn1==0.4.2
pyasn1-modules==0.2.1
pykml==0.1.3
python-dateutil==2.7.0
python-geoip==1.2
python-geoip-geolite2==2015.303
python-mimeparse==1.6.0
pytz==2017.2                        # Timezone support
requests==2.18.4
rsa==3.4.2
sendwithus==5.0.0
shortuuid==0.5.0
simplejson==3.13.2
six==1.11.0
urllib3==1.22
-e git://github.com/joram/stravalib.git#egg=stravalib
-e git://github.com/wckd/python-yr.git#egg=python-yr
